Corrected typehint

+++ primesieve/__init__.pyi
@@ -1,6 +1,6 @@
 import numpy as np
 import numpy.typing as npt
 
-def sieve(limit: int) -> npt.NDArray[np.int32]:
+def sieve(limit: int) -> npt.NDArray[np.int64]:
     """Return all prime numbers up to a given limit using Sieve of Eratosthenes."""
     ...
